fix: 소소한 조회 최적화
- 게시글 조회시 member와 함꼐 조회
- find보다는 exist를 사용하여 명시적

+++ src/main/java/com/kakaotechbootcamp/community/application/post/service/PostLikeService.java
@@ -28,8 +28,10 @@ public class PostLikeService {
      */
     @Transactional
     public void likePost(Long postId, Long memberId) {
-        Post post = findPostById(postId);
-        Member member = findMemberById(memberId);
+        Post post = postRepository.findByIdWithMember(postId)
+                .orElseThrow(() -> new CustomException(PostErrorCode.POST_NOT_FOUND));
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new CustomException(MemberErrorCode.USER_NOT_FOUND));
 
         postLikePolicy.validateCanLike(postId, memberId);
 
@@ -42,20 +44,13 @@ public class PostLikeService {
      */
     @Transactional
     public void unlikePost(Long postId, Long memberId) {
-        findPostById(postId); // 게시글 존재 여부 확인
+        if(!postRepository.existsById(postId)){
+            throw new CustomException(PostErrorCode.POST_NOT_FOUND);
+        }
         postLikePolicy.validateCanUnlike(postId, memberId);
 
         postLikeRepository.deleteByPostIdAndMemberId(postId, memberId);
         postRepository.decrementLikeCount(postId);
     }
 
-    private Post findPostById(Long postId) {
-        return postRepository.findByIdWithMember(postId)
-                .orElseThrow(() -> new CustomException(PostErrorCode.POST_NOT_FOUND));
-    }
-
-    private Member findMemberById(Long memberId) {
-        return memberRepository.findById(memberId)
-                .orElseThrow(() -> new CustomException(MemberErrorCode.USER_NOT_FOUND));
-    }
 }
